refactor: rename back to modulo_controllers
